Move IcebergParquetReader declaration into header
